+++ packages/tailwind-config/postcss.config.cjs
@@ -1,8 +1,1 @@
-module.exports = {
-  plugins: [
-    require('postcss-import'),
-    require('postcss-preset-env'),
-    // require('tailwindcss'),
-    require('autoprefixer'),
-  ],
-};
+module.exports = require('../../postcss.config.cjs');